@@ -25,10 +25,10 @@
 package com.heinrichreimersoftware.materialintro.demo;
 
 import android.content.Intent;
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.heinrichreimersoftware.materialintro.demo.databinding.ActivitySplashBinding;
 
