@@ -24,7 +24,7 @@
 
 package com.heinrichreimersoftware.materialintro.app;
 
-import android.support.annotation.IntDef;
+import androidx.annotation.IntDef;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
