@@ -24,7 +24,7 @@
 
 package com.heinrichreimersoftware.materialintro.app;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 
 public class SlideFragment extends Fragment implements IntroNavigation {
