@@ -24,7 +24,7 @@
 
 package com.heinrichreimersoftware.materialintro.slide;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 public interface RestorableSlide extends Slide {
     void setFragment(Fragment fragment);
