@@ -18,7 +18,7 @@ package com.heinrichreimersoftware.materialintro.util;
 
 import android.content.Context;
 import android.os.Build;
-import android.support.v4.view.animation.FastOutSlowInInterpolator;
+import androidx.interpolator.view.animation.FastOutSlowInInterpolator;
 import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
