@@ -1,9 +1,9 @@
 package com.heinrichreimersoftware.materialintro.view.parallax;
 
 import android.os.Bundle;
-import android.support.annotation.FloatRange;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.FloatRange;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 
 import com.heinrichreimersoftware.materialintro.view.parallax.util.ParallaxUtil;
