@@ -24,7 +24,7 @@
 
 package com.heinrichreimersoftware.materialintro.demo;
 
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.os.Handler;
 import android.view.LayoutInflater;
