@@ -27,7 +27,7 @@ package com.heinrichreimersoftware.materialintro.demo;
 import android.Manifest;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.Snackbar;
+import com.google.android.material.snackbar.Snackbar;
 import android.view.Gravity;
 import android.view.View;
 import android.widget.Toast;
